+++ BookClub/forms/__init__.py
@@ -2,3 +2,5 @@ from .authentication import *
 from .club import *
 from .meeting import *
 from .user_forms import *
+from .booklist_forms import *
+from .pagination_forms import *
