+++ BookClub/views/booklists.py
@@ -1,10 +1,11 @@
+from django.contrib import messages
 from django.shortcuts import get_object_or_404
 from django.contrib.auth.mixins import LoginRequiredMixin
 
-from django.views.generic.list import ListView
+from django.views.generic import ListView, CreateView
 
-from BookClub.models.user import User
-from BookClub.models.booklist import BookList
+from BookClub.models import User, BookList
+from BookClub.forms import CreateBookListForm
 
 class BooklistListView(ListView):
     http_method_names = ['get']
@@ -24,3 +25,18 @@ class BooklistListView(ListView):
         context['creator'] = creator
         context['self'] = self.request.user == creator
         return context
+
+class CreateBookListView(LoginRequiredMixin, CreateView):
+    template_name = 'create_booklist.html'
+    model = BookList
+    form_class = CreateBookListForm
+
+    def form_valid(self, form):
+        form.instance.creator = self.request.user
+        response = super().form_valid(form)
+        messages.success(self.request, ('Successfully created a new booklist!'))
+        return response
+
+    def form_invalid(self, form):
+        messages.add_message(self.request, messages.ERROR, "The data provided was invalid!")
+        return super().form_invalid(form)
